+++ pkg/cli/completion.go
@@ -1,7 +1,10 @@
 package cli
 
 import (
+	"encoding/json"
 	"fmt"
+	"io"
+	"os"
 	"strconv"
 	"strings"
 
@@ -25,60 +28,45 @@ func BashCompletionScript(command string) string {
 	return fmt.Sprintf(bashCompletionTemplate, command)
 }
 
-// FormatCompletionSuggestions takes pre-filtered completion suggestions and
-// returns a formatted string ready to pass back to the shell during a
-// completion request. The description and addition suffixes are dropped when
-// only one option is available.
-func FormatCompletionSuggestions(width int, suggestions []option.Description) string {
-	var s strings.Builder
-	if len(suggestions) == 1 {
-		v := suggestions[0].Option
-		v = strings.Split(v, " ")[0]
-		fmt.Fprintf(&s, "%v\n", v)
-	} else {
-		fmt.Fprintf(&s, "%v", option.FormatCompletion(width, suggestions))
-	}
-
-	return s.String()
+// DefaultCompletion implements a default completion for a given option field,
+// and can be used a fallback by command completion handlers. It handles
+// specific cases based on the option field type, and simulates default shell
+// behavior (filename completion) for string types.
+func DefaultCompletion(opt *option.T, w string) []string {
+	return DefaultFilenameCompletion(opt, w)
 }
 
-// DefaultCompletion acts like the shell default completion and suggests file
-// and folder names under the current directory. It is used by default when the
-// command does not implement a specific completion handler, and should be used
-// from the command completion handler when no other completion logic is
-// suitable.
-func DefaultCompletion(w string) []option.Description {
-	var r []option.Description
+// DefaultFilenameCompletion implements a default filename-based completion,
+// similar to the default shell completion.
+func DefaultFilenameCompletion(opt *option.T, w string) (r []string) {
 	files, err := dir.Glob(fmt.Sprintf("%v*", w))
 	if err == nil {
 		if len(files) == 1 && strings.HasSuffix(files[0], string(fileutils.Separator)) {
-			return DefaultCompletion(files[0])
+			return DefaultFilenameCompletion(opt, files[0])
 		}
 		for _, f := range files {
-			r = append(r, option.Description{Option: f})
+			r = append(r, f)
 		}
 	}
-	return r
+	return
 }
 
-// FilepathCompletion implements a custom filepath completion scheme, matching
-// the provided pattern if possible. If the pattern does not yield any result,
-// or if none of the suggestions match the partial argument being completed,
-// then if folds back to the default filename completion.
-func FilepathCompletion(pattern string, w string) []option.Description {
+// MatchingFileCompletion implements a custom filepath completion scheme,
+// matching the provided pattern if possible, using default filename completion
+// as a fallback.
+func MatchingFilenameCompletion(opt *option.T, pattern string, w string) (r []string) {
 	files, err := dir.Glob(pattern)
 	if err == nil {
-		var r []option.Description
 		for _, f := range files {
 			if strings.HasPrefix(f, w) {
-				r = append(r, option.Description{Option: f})
+				r = append(r, f)
 			}
 		}
 		if len(r) > 0 {
 			return r
 		}
 	}
-	return DefaultCompletion(w)
+	return DefaultFilenameCompletion(opt, w)
 }
 
 // ---
@@ -87,6 +75,10 @@ func (cmd *Command) handleCompletionRequest() bool {
 	i, w := checkCompletionRequest(cmd.ProcessEnv)
 	if i == 0 {
 		return false
+	}
+
+	if w == "--" {
+		w = ""
 	}
 
 	// Trim process arguments past the completion point
@@ -99,28 +91,30 @@ func (cmd *Command) handleCompletionRequest() bool {
 	}
 
 	var comp = cmd.opts.GetCompletion(args, w)
-	if comp.Opt != nil {
-		comp.OptValues = cmd.complete(comp.Opt, w)
+	if comp.OptRef != nil {
+		comp.OptValues = cmd.complete(comp.OptRef, w)
 	}
-	if comp.Arg != nil {
-		comp.ArgValues = cmd.complete(comp.Arg, w)
+	if comp.ArgRef != nil {
+		comp.ArgValues = cmd.complete(comp.ArgRef, w)
 	}
 
 	for _, o := range comp.Options {
-		if strings.HasPrefix(o.Option, w) {
+		if strings.HasPrefix(o, w) {
 			cmd.Suggestions = append(cmd.Suggestions, o)
 		}
 	}
 	for _, o := range comp.OptValues {
-		if strings.HasPrefix(o.Option, w) {
+		if strings.HasPrefix(o, w) {
 			cmd.Suggestions = append(cmd.Suggestions, o)
 		}
 	}
 	for _, o := range comp.ArgValues {
-		if strings.HasPrefix(o.Option, w) {
+		if strings.HasPrefix(o, w) {
 			cmd.Suggestions = append(cmd.Suggestions, o)
 		}
 	}
+
+	dumpCompletionRequest(comp)
 
 	return true
 }
@@ -137,9 +131,29 @@ func checkCompletionRequest(env map[string]string) (index int, word string) {
 	return
 }
 
-func (cmd *Command) complete(opt *option.T, word string) []option.Description {
+func dumpCompletionRequest(comp option.Completion) {
+	if s, ok := os.LookupEnv("COMPLETION_DEBUG_OUTPUT"); ok && s != "" {
+		type obj = map[string]interface{}
+		var v = obj{
+			"args": os.Args,
+			"env": obj{
+				"COMP_INDEX": os.Getenv("COMP_INDEX"),
+				"COMP_WORD":  os.Getenv("COMP_WORD"),
+			},
+			"comp": comp,
+		}
+
+		fileutils.WriteFile(s, func(w io.Writer) error {
+			var e = json.NewEncoder(w)
+			e.SetIndent("", "    ")
+			return e.Encode(v)
+		})
+	}
+}
+
+func (cmd *Command) complete(opt *option.T, word string) []string {
 	if handler, ok := cmd.Handler.(CompletionHandler); ok {
 		return handler.Complete(opt, word)
 	}
-	return DefaultCompletion(word)
+	return DefaultCompletion(opt, word)
 }